pyproject.toml
[tool.poetry]
name = "dj-auth"
version = "0.1.0"
description = ""
authors = ["makkahwi <suhaibahmadai@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
